Advent of Code 2019 - Day 2 * Cleanup of early Intcode examples

+++ src/main/java/net/spizzer/aoc2019/intcode/IntcodeComputer.java
@@ -111,7 +111,7 @@ public class IntcodeComputer {
         return output.get(output.size() - 1);
     }
 
-    public long getLastIntOutput() {
+    public int getLastIntOutput() {
         return Math.toIntExact(output.get(output.size() - 1));
     }
